+++ packages/@react-native-windows/find-repo-root/CHANGELOG.md
@@ -1,8 +1,18 @@
 # Change Log - @react-native-windows/find-repo-root
 
-This log was last generated on Wed, 24 Jul 2024 05:24:36 GMT and should not be manually modified.
+This log was last generated on Thu, 08 Aug 2024 05:16:50 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.0.0-canary.84
+
+Thu, 08 Aug 2024 05:16:50 GMT
+
+### Changes
+
+- Bump @react-native-windows/fs to v0.0.0-canary.55
+- Bump @rnw-scripts/eslint-config to v1.2.26
+- Bump @rnw-scripts/just-task to v2.3.43
 
 ## 0.0.0-canary.83